Training refactoring and logging
