+++ training/rootnav2/loss/__init__.py
@@ -1,19 +1,13 @@
 import copy
-import logging
 import functools
 
 from rootnav2.loss.loss import cross_entropy2d
-
-
-
-logger = logging.getLogger('rootnav2')
 
 key2loss = {'cross_entropy': cross_entropy2d,
             }
 
 def get_loss_function(cfg):
     if cfg['training']['loss'] is None:
-        logger.info("Using default cross entropy loss")
         return cross_entropy2d
 
     else:
@@ -24,6 +18,4 @@ def get_loss_function(cfg):
         if loss_name not in key2loss:
             raise NotImplementedError('Loss {} not implemented'.format(loss_name))
 
-        logger.info('Using {} with {} params'.format(loss_name, 
-                                                     loss_params))
         return functools.partial(key2loss[loss_name], **loss_params)
